refactor: split transition file into stf and preprocess
