+++ crates/asm/logs/src/lib.rs
@@ -1,3 +1,8 @@
+//! # ASM Log Types
+//!
+//! This crate provides structured log types for the Anchor State Machine (ASM) in the Strata
+//! protocol. It defines various log entry types that capture important events within the system.
+
 pub mod asm_stf;
 pub mod checkpoint;
 pub mod constants;
