+++ crates/asm/stf/src/preprocess.rs
@@ -1,0 +1,92 @@
+//! The `asm_stf` crate implements the core Anchor State Machine state transition function (STF). It
+//! glues together block‐level validation, a set of pluggable subprotocols, and the global chain
+//! view into a single deterministic state transition.
+
+use std::collections::BTreeMap;
+
+use bitcoin::{block::Block, params::Params};
+use strata_asm_common::{AnchorState, AsmError, AsmResult, AsmSpec};
+
+use crate::{
+    manager::SubprotoManager,
+    stage::{PreProcessStage, SubprotoLoaderStage},
+    tx_filter::group_txs_by_subprotocol,
+    types::AsmPreProcessOutput,
+};
+
+/// Pre-processes a Bitcoin block for the Anchor State Machine (ASM) state transition.
+///
+/// This function performs the initial phase of ASM block processing that must be completed
+/// before running the main ASM state transition function [`crate::asm_stf`].
+///
+/// This function performs the initial phase of ASM block processing, which validates block header
+/// continuity against the current chain state, filters and groups transactions by subprotocol,
+/// and pre-processes transactions to collect the [`AuxRequest`](strata_asm_common::AuxRequest)
+///
+/// # Arguments
+///
+/// * `pre_state` - The current anchor state containing chain view and subprotocol states
+/// * `block` - The Bitcoin block to pre-process
+///
+/// # Returns
+///
+/// Returns an `AsmResult` containing:
+/// - `AsmPreProcessOutput` with filtered transactions and auxiliary requests on success
+/// - `AsmError` if validation fails or pre-processing encounters an error
+///
+/// # Errors
+///
+/// This function will return an error if:
+/// - The block header fails PoW continuity validation
+/// - Subprotocol loading or pre-processing fails
+///
+/// # Type Parameters
+///
+/// * `S` - The ASM specification type that defines magic bytes and subprotocol behavior
+/// * `'b` - Lifetime parameter tied to the input block reference
+pub fn pre_process_asm<'b, S: AsmSpec>(
+    pre_state: &AnchorState,
+    block: &'b Block,
+) -> AsmResult<AsmPreProcessOutput<'b>> {
+    // 1. Validate and update PoW header continuity for the new block.
+    // This ensures the block header follows proper Bitcoin consensus rules and chain continuity.
+    let mut pow_state = pre_state.chain_view.pow_state.clone();
+    pow_state
+        .check_and_update_continuity(&block.header, &Params::MAINNET)
+        .map_err(AsmError::InvalidL1Header)?;
+
+    // 2. Filter and group transactions by subprotocol based on magic bytes.
+    // Only transactions relevant to registered subprotocols are processed further.
+    let grouped_relevant_txs = group_txs_by_subprotocol(S::MAGIC_BYTES, &block.txdata);
+
+    let mut manager = SubprotoManager::new();
+
+    // 3. LOAD: Initialize each subprotocol in the subproto manager.
+    // We use empty auxpayload in the loader stage as no auxiliary data is needed during loading.
+    let aux = BTreeMap::new();
+    let mut loader_stage = SubprotoLoaderStage::new(pre_state, &mut manager, &aux);
+    S::call_subprotocols(&mut loader_stage);
+
+    // 4. PROCESS: Feed each subprotocol its filtered transactions for pre-processing.
+    // This stage extracts auxiliary requests that will be needed for the main STF execution.
+    let mut pre_process_stage =
+        PreProcessStage::new(&grouped_relevant_txs, &mut manager, pre_state);
+    S::call_subprotocols(&mut pre_process_stage);
+
+    // 5. Flatten the grouped transactions back into a single collection.
+    // The grouping was needed for per-subprotocol processing, but the output needs a flat list.
+    let relevant_txs = grouped_relevant_txs
+        .into_iter()
+        .flat_map(|(_k, vec)| vec)
+        .collect();
+
+    // 6. Export auxiliary requests collected during pre-processing.
+    // These requests will be fulfilled before running the main ASM state transition.
+    let aux_requests = manager.export_aux_requests();
+    let output = AsmPreProcessOutput {
+        txs: relevant_txs,
+        aux_requests,
+    };
+
+    Ok(output)
+}
